Add "readonly" attribute to HTML input element
This ensures that the value in the input field can only be changed via
the datepicker. Without this attribute, users may click into the input
field and add arbitrary text by typing on their keyboard. This should
almost never be what one (as a developer) wants when using a datepicker
widget.

Note that this is completely orthogonal to the datepicker's
`enableOnReadonly` and to SurveyJS's `isReadOnly` attributes.